[CHG] Using s8 instead of int for the page variable in lesson 2

+++ examples/lesson2/source/main.c
@@ -15,7 +15,7 @@
 
 int main() {
     u32 wpaddown;
-    int page = 0;
+    s8 page = 0;
 
     // Font texture
     GRRLIB_texImg *text_font1 = GRRLIB_LoadTexture(font1);
